Log actual and expected checksums on mismatch
Without this patch, updating Jenkins plugins is difficult because it
isn't clear what the actual checksum of the downloaded package is.  The
end user may discover the checksum by running puppet in debug mode,
discovering the curl command used to download the jenkins plugin, run
the command manually, then compute the checksum manually.

This patch helps the user by displaying the actual and expected
checksums when comparing them.  This allows the user to copy and paste
the actual checksum when updating hiera with the expected checksum.