Rewrite artifactory_sha1 function with puppet v4 api
Old v3 API function is removed, and new `archive::artifactory_checksum`
function added.  Puppet 4 API functions should not have issues with
environment isolation.

This is a breaking change.  The function is renamed and has been
upgraded. It can now return md5 or sha256 checksums as well as sha1.

The function now expects the artifact URL and will convert to the API
url internally. (This was previously done in the calling puppet code.)